% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooper.R
\name{cooper}
\alias{cooper}
\title{Cooper from EpiPets}
\usage{
cooper(dataframe, var)
}
\arguments{
\item{dataframe}{A data frame.}

\item{var}{A categorical variable.}
}
\value{
A nicely formatted count of a single variable's levels in n (\%) form.
}
\description{
A quick function for clean counts and percentages of categorical variables.
}
\examples{
cooper(mtcars, cyl)

}
