% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kissie.R
\name{kissie}
\alias{kissie}
\title{Kissie From EpiPets}
\usage{
kissie(dataframe, outcome, exposure)
}
\arguments{
\item{dataframe}{A dataframe.}

\item{outcome}{A Binary Outcome Variable}

\item{exposure}{A Categorical Exposure Variable}
}
\value{
A table with n x 2 cells with odds ratios and p-values matched to their corresponding rows. Ensure reference levels are properly designated before using.
}
\description{
Takes tabby() and loggy() and makes them kiss! Helpful for a manucript's table 2 with matching logistic regression output to tabulated n(\%) cell values. Can be used on its own but is internally built to be seamlessly compatible with flextable() functions.
}
\examples{

data(mtcars)
mtcars$am <- factor(mtcars$am)
mtcars$cyl <- factor(mtcars$cyl)
kissie(mtcars, am, cyl)

}
