% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggy.R
\name{loggy}
\alias{loggy}
\title{Loggy (Simple Logistic Regression) from EpiPets}
\usage{
loggy(dataframe, y, x)
}
\arguments{
\item{dataframe}{A data frame.}

\item{y}{A binary dependent variable.}

\item{x}{An independent variable.}
}
\value{
A data frame containing the odds ratio, 95\% confidence interval, and p-value for each model term. For categorical predictors, ensure the desired reference level is set before fitting the model.
}
\description{
I expect you to know how logistic regression works :P
(jk thats why i made this function)
Function to Produce ORs, 95\% CI, and P-Values from logistic regression.
}
\examples{
# Binary outcome with continuous predictor
loggy(mtcars, am, wt)

# Binary outcome with binary predictor
loggy(mtcars, am, vs)

}
