% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabby.R
\name{tabby}
\alias{tabby}
\title{Tabby (Tabulate) from EpiPets}
\usage{
tabby(dataframe, var, group)
}
\arguments{
\item{dataframe}{A data frame.}

\item{var}{A chosen categorical variable, for example, race.}

\item{group}{A chosen categorical variable by which to group counts, for example, Treatment Group.}
}
\value{
A dataframe with levels of "var" as rows, and levels of "group" as columns. Cells contain n (\%), where cell percentages are calculated as part of column totals.
}
\description{
Function for tabulating counts of one variable grouped by another in n(\%) form.
}
\examples{
tabby(mtcars, cyl, gear)

}
