% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookie.R
\name{lookie}
\alias{lookie}
\title{Lookie (Lookup and Summarize)}
\usage{
lookie(dataframe, var, group, match = c("exact", "prefix"))
}
\arguments{
\item{dataframe}{A data frame.}

\item{var}{An unquoted column-name (variable) prefix used to identify columns to search.}

\item{group}{A named list of vectors containing values to search for.}

\item{match}{A matching argument that defaults to exact term matching. Specify match = "prefix" to use prefix-based matching of values. When using exact matching, a search for A15 will return false for a value A150 or A156. When using prefix matching, searching for A15 will return true for a value A15, A159, A1555, and so on.}
}
\value{
A tibble containing each group and its count and percentage formatted as n (\%), followed by a total row. Please pay attention to the output inherent to the structure of this function. If you utilize multiple values across multiple columns, percent values in the output will not sum to 100\%.
}
\description{
A useful tool to search for values or groups of values across one column or multiple columns. I built this for initial analysis with multiple columns of diagnosis codes, but it has various applications.
}
\examples{
# Aquire a Dataset with lots of repeating categorical values across multiple columns.
store <- data.frame(
id = 1:100,
produce_mon = sample(c("apple", "banana", "kiwi", "brocolli", "potato", "lettuce"),
100, replace = TRUE),
produce_tue = sample(c("apple", "banana", "kiwi", "brocolli", "potato", "lettuce"),
100, replace = TRUE),
produce_wed = sample(c("apple", "banana", "kiwi", "brocolli", "potato", "lettuce"),
100, replace = TRUE)
)
# Assign desired groups to count by.
fruits <- c("apple", "banana", "kiwi")
veggies <- c("brocolli", "potato", "lettuce")

# Bring your groups into one list for lookie() to use.
shopping_list <- list(
 fruits = fruits,
 veggies = veggies
)

lookie(store, produce, shopping_list)

}
